Air fly project
The project is a website for booking flight tickets to various destinations. The logic is implemented using C#. For the client-side interface, I utilized Windows Forms. 

I chose this project because it is complex and challenging, requiring careful consideration of multiple factors, such as the number of passengers, their ages, and class selections. Additionally, the project emphasizes the principles of object-oriented programming, allowing for a structured and efficient approach to managing data and functionality. This experience not only enhanced my technical skills but also improved my ability to solve problems and think critically in a software development context.